clean up submit model
