+++ app/concerns/auto_check_sender.rb
@@ -1,4 +1,4 @@
-class CheckServer::Uploader
+class AutoCheckSender
 	
 	def self.enabled?
 		ENV['CHECK_SERVER_URL'].present? && ENV['CHECK_SERVER_SECRET'].present?
