+++ app/controllers/profile_controller.rb
@@ -8,7 +8,7 @@ class ProfileController < ApplicationController
 	
 	def feedback
 		submit = Submit.find(params[:submit_id])
-		@formatted_feedback = submit.check_feedback_formatted
+		@formatted_feedback = submit.formatted_auto_feedback
 		respond_to do |format|
 			format.js do
 				render 'feedback'
