+++ app/views/grading/show.html.erb
@@ -49,7 +49,7 @@
 
 <div class="col-7" style="overflow-y:scroll; min-height:0; box-shadow:#cccccc 2px 0px 5px inset;max-height:100%">
 	
-	<% if @files.present? || @submit.has_feedback? %>
+	<% if @files.present? || @submit.has_auto_feedback? %>
 	<ul class="nav nav-tabs" id="files" role="tablist">
 		<% first = true %>
 		<% @files.each_key do |filename| %>
@@ -57,7 +57,7 @@
 			<a class="nav-link <%= (first = false) || 'active' if first %>" id="home-tab" data-toggle="tab" href="#file-<%= filename.parameterize %>" role="tab"><%= filename %></a>
 		</li>
 		<% end if @files.present? %>
-		<% if @submit.has_feedback? %>
+		<% if @submit.has_auto_feedback? %>
 		<li class="nav-item">
 			<a class="nav-link <%= (first = false) || 'active' if first %>" id="home-tab" data-toggle="tab" href="#file-check" role="tab">check</a>
 		</li>
@@ -70,8 +70,8 @@
 			<%= link_to 'd', grading_download_path(grading_submit_id: @submit.id, filename: filename), class: 'pull-right' %><%= CodeRay.scan(contents, CodeRay::FileType.fetch(filename, :text)).div(:line_numbers => :inline).html_safe %>
 		</div>
 		<% end if @files.present? %>
-		<% if @submit.has_feedback? %>
-		<div class="tab-pane fade <%= (first = false) || 'show active' if first %>" id="file-check" role="tabpanel" aria-labelledby="home-tab"><code><pre><%= @submit.check_feedback_formatted %></code></pre></div>
+		<% if @submit.has_auto_feedback? %>
+		<div class="tab-pane fade <%= (first = false) || 'show active' if first %>" id="file-check" role="tabpanel" aria-labelledby="home-tab"><code><pre><%= @submit.formatted_auto_feedback %></code></pre></div>
 		<% end %>
 	</div>
 	<% end %>
