+++ app/views/items/_submit.html.erb
@@ -14,7 +14,7 @@ URL: <%= link_to submit.url, submit.url %>
 	</ul>
 </p>
 <% end %>
-<% if submit.has_feedback? %>
+<% if submit.has_auto_feedback? %>
 <%= link_to t("check_results"), profile_feedback_path(submit), class:"btn btn-secondary", data: { toggle:'modal', target:'#fb-modal', submit: submit.id } %>
 <% end %>
