+++ app/views/submits/_info.html.erb
@@ -1,3 +1,3 @@
 <p class="mb-0">
-	Last submitted: <%= @submit.last_submitted %> <%= "via login #{@submit.used_login}" if @submit.used_login %>
+	Last submitted: <%= some_time_or_never(@submit.submitted_at) %> <%= "via login #{@submit.used_login}" if @submit.used_login %>
 </p>
